requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
langchain-pinecone
sentence-transformers
pypdf
python-dotenv
pinecone-client
-e .
